backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose
sentence-transformers
langchain
langchain-groq
pydantic
